+++ ThirdPartyAdapters/bidmachine/bidmachine/src/test/kotlin/com/google/ads/mediation/bidmachine/BidMachineNativeAdTest.kt
@@ -15,14 +15,17 @@
 package com.google.ads.mediation.bidmachine
 
 import android.content.Context
+import androidx.core.os.bundleOf
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.ads.mediation.adaptertestkit.AdErrorMatcher
 import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_BID_RESPONSE
+import com.google.ads.mediation.adaptertestkit.AdapterTestKitConstants.TEST_PLACEMENT_ID
 import com.google.ads.mediation.adaptertestkit.createMediationNativeAdConfiguration
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ADAPTER_ERROR_DOMAIN
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ERROR_CODE_AD_REQUEST_EXPIRED
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.ERROR_MSG_AD_REQUEST_EXPIRED
+import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.PLACEMENT_ID_KEY
 import com.google.ads.mediation.bidmachine.BidMachineMediationAdapter.Companion.SDK_ERROR_DOMAIN
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.mediation.MediationAdLoadCallback
@@ -69,18 +72,36 @@ class BidMachineNativeAdTest {
 
   @Before
   fun setUp() {
+    val serverParams = bundleOf(PLACEMENT_ID_KEY to TEST_PLACEMENT_ID)
     val adConfiguration =
-      createMediationNativeAdConfiguration(context = context, bidResponse = TEST_BID_RESPONSE)
+      createMediationNativeAdConfiguration(
+        context = context,
+        bidResponse = TEST_BID_RESPONSE,
+        serverParameters = serverParams,
+      )
     BidMachineNativeAd.newInstance(adConfiguration, mockAdLoadCallback).onSuccess {
       bidMachineNativeAd = it
     }
   }
 
   @Test
-  fun loadAd_invokesBidMachineRequest() {
+  fun loadWaterfallAd_invokesBidMachineRequest() {
     val mockNativeRequestBuilder = configureNativeRequestBuilder()
 
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadWaterfallAd(mockNativeAd)
+
+    verify(mockNativeAd).setListener(eq(bidMachineNativeAd))
+    verify(mockNativeRequestBuilder, never()).setBidPayload(any())
+    verify(mockNativeRequestBuilder).setPlacementId(eq(TEST_PLACEMENT_ID))
+    verify(mockNativeRequestBuilder).setListener(eq(bidMachineNativeAd))
+    verify(mockNativeRequest).request(eq(context))
+  }
+
+  @Test
+  fun loadRtbAd_invokesBidMachineRequest() {
+    val mockNativeRequestBuilder = configureNativeRequestBuilder()
+
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
 
     verify(mockNativeAd).setListener(eq(bidMachineNativeAd))
     verify(mockNativeRequestBuilder).setBidPayload(eq(TEST_BID_RESPONSE))
@@ -90,7 +111,7 @@ class BidMachineNativeAdTest {
 
   @Test
   fun onRequestSuccess_invokesBannerViewLoad() {
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
 
     bidMachineNativeAd.onRequestSuccess(mockNativeRequest, mock())
 
@@ -102,7 +123,7 @@ class BidMachineNativeAdTest {
     whenever(mockNativeRequest.isExpired) doReturn true
     val expectedAdError =
       AdError(ERROR_CODE_AD_REQUEST_EXPIRED, ERROR_MSG_AD_REQUEST_EXPIRED, ADAPTER_ERROR_DOMAIN)
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
 
     bidMachineNativeAd.onRequestSuccess(mockNativeRequest, mock())
 
@@ -136,7 +157,7 @@ class BidMachineNativeAdTest {
   @Test
   fun onAdLoaded_invokesOnSuccess() {
     configureNativeRequestBuilder()
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
 
     bidMachineNativeAd.onAdLoaded(mockNativeAd)
 
@@ -147,7 +168,7 @@ class BidMachineNativeAdTest {
   fun onAdLoadFailed_invokesOnFailure() {
     val bMError = BMError.AlreadyShown
     val expectedAdError = AdError(bMError.code, bMError.message, SDK_ERROR_DOMAIN)
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
 
     bidMachineNativeAd.onAdLoadFailed(mockNativeAd, bMError)
 
@@ -158,7 +179,7 @@ class BidMachineNativeAdTest {
   @Test
   fun onAdImpression_invokesReportAdImpression() {
     configureNativeRequestBuilder()
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
     bidMachineNativeAd.onAdLoaded(mockNativeAd)
 
     bidMachineNativeAd.onAdImpression(mockNativeAd)
@@ -169,7 +190,7 @@ class BidMachineNativeAdTest {
   @Test
   fun onAdClicked_invokesOnAdOpenedOnAdLeftApplicationAndReportAdClicked() {
     configureNativeRequestBuilder()
-    bidMachineNativeAd.loadAd(mockNativeAd)
+    bidMachineNativeAd.loadRtbAd(mockNativeAd)
     bidMachineNativeAd.onAdLoaded(mockNativeAd)
 
     bidMachineNativeAd.onAdClicked(mockNativeAd)
@@ -192,6 +213,7 @@ class BidMachineNativeAdTest {
   private fun configureNativeRequestBuilder(): NativeRequest.Builder {
     val mockNativeRequestBuilder =
       mock<NativeRequest.Builder> {
+        on { setPlacementId(TEST_PLACEMENT_ID) } doReturn it
         on { setBidPayload(eq(TEST_BID_RESPONSE)) } doReturn it
         on { setListener(any()) } doReturn it
         on { build() } doReturn mockNativeRequest